Change hyphenInterval type from Short to Byte

+++ library/encoding-base32/src/commonMain/kotlin/io/matthewnelson/encoding/base32/Base32.kt
@@ -82,7 +82,7 @@ public sealed class Base32(config: EncoderDecoder.Config): EncoderDecoder(config
             @JvmField
             public val encodeToLowercase: Boolean,
             @JvmField
-            public val hyphenInterval: Short,
+            public val hyphenInterval: Byte,
             @JvmField
             public val checkByte: Byte?,
         ): EncoderDecoder.Config(isLenient, paddingByte = null) {
